+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add tile option to specify filename of a multifile gist.
 
 
 0.1 (2017-06-04)