+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-0.2 (unreleased)
+0.2 (2017-07-28)
 ----------------
 
 - Add tile option to specify filename of a multifile gist.